@@ -6,7 +6,7 @@ authors:
 homepage: https://github.com/fgcan10/chart_components
 
 environment:
-  sdk: ">=2.12.0-0 <3.0.0"
+  sdk: '>=3.1.2 <4.0.0'
 
 dependencies:
   flutter:
